+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Better error handling in `check` method with better error messages.
 - Handle error if the `dd --version` format isn't as expected.
 
+### Changed
+- Use `Vec<String>` in options instead of `Vec<(String, String)>` for better performance and greater code.
+
 ## [0.1.0] - 2025-02-04
 
 ### Added